+++ __tests__/commands/check.spec.js
